More options to scripts

+++ clean.sh
@@ -2,5 +2,8 @@
 
 # Removes training and testing temp files
 rm data/models -r
-rm data/vectorizers -r
 rm data/predictions -r
+
+if [ $1 != "--keep-vectorizers" ]; then
+    rm data/vectorizers -r
+fi
